ZDBSP: remove a dubious cast from `unsigned int` to `long`
See https://github.com/ZDoom/gzdoom/commit/ab8711b57ade09349fa29cb8a19d552610fafc8e.

+++ zdbsp/src/utility.cpp
@@ -290,7 +290,7 @@ void FNodeBuilder::FindPolyContainers(TArray<FPolyStart>& spots, TArray<FPolySta
 
 						if (dist < closestdist && dist >= 0) {
 							closestdist = dist;
-							closestseg = (long)j;
+							closestseg = j;
 						}
 					}
 				}
